refactor: reservation 개행 수정 및 폴더명 수정
