+++ src/components/reservation/ReservationCom.js
@@ -1,8 +1,9 @@
 import React, { useState, forwardRef } from "react";
 import DatePicker from "react-datepicker";
+import styled, { css } from "styled-components";
+
 import "react-datepicker/dist/react-datepicker.css";
 import { ko } from "date-fns/esm/locale";
-import styled, { css } from "styled-components";
 import { set } from "date-fns";
 
 //test라우터
